Convert FormEdit store into a Zustand slice
